Change ElboDeriv to use the linear approximation
