+++ test/runtests.jl
@@ -9,6 +9,7 @@ using Transform
 import Synthetic
 
 include(joinpath(Pkg.dir("Celeste"), "test", "test_misc.jl"))
+include(joinpath(Pkg.dir("Celeste"), "test", "test_wcs.jl"))
 include(joinpath(Pkg.dir("Celeste"), "test", "test_kl.jl"))
 include(joinpath(Pkg.dir("Celeste"), "test", "test_constraints.jl"))
 include(joinpath(Pkg.dir("Celeste"), "test", "test_elbo_values.jl"))
